refactor(ingest): decompose processing/ingest/metadata.py
metadata.py was 2151 lines with five carve-outs on its size cap, each one a
correctness fix that had to argue for its lines (#888, #899, #906, #934, #961,
then #1104/#1113's eight rounds). #958 named the problem — the ratchet taxing
correctness work on a module nobody had time to split — and filed #1042.

Split into seven siblings, each one a cluster the carve-outs kept landing on:

  metadata_sql (60)          identifier validation and quoting
  metadata_quality (190)     the four weighted score dimensions
  metadata_projection (266)  geom_4326, its GIST index, the reader grant
  metadata_attributes (352)  attribute-metadata rows and their inference
  metadata_mercator (361)    the clip, the 0..360 shift, the CRS predicates
  metadata_geometry (424)    constructing geom, laundering column names
  metadata_extent (633)      reading metadata back off a landed table

metadata_sql holds only what every other module needs and needs nothing back,
which is what keeps the other six from importing each other.

Two seams differ from the issue's sketch. The SQL CRS predicates stay with the
clip rather than becoming their own module: `_GEOGRAPHIC_SRTEXT_RE` and
`_DEGREE_UNIT_SRTEXT_RE` have exactly one caller each, and #961's standing
obligation is that the anchored gate and the degenerate-envelope floor
deliberately disagree — putting the two halves of that disagreement in separate
files would make it unreadable. And the extent producers keep company with the
four introspection helpers, because `extract_metadata`'s CTE consolidates those
same four and falls back to calling them one at a time, so the CTE and the
helpers have to be legible together.

No behavior change. All 49 top-level symbols are byte-identical to their
pre-split text (verified by AST segment comparison against the parent commit),
and metadata.py re-exports every one of them, so existing importers and the
`mock.patch("app.processing.ingest.metadata.<name>")` targets keep resolving.
The re-exports carry `# noqa: F401` plus an explicit `__all__` so ruff's
autofix cannot strip them; `ruff check --fix --diff` produces no diff.

Two references had to move to the module a call physically lives in, because
the façade name is not what the code reads:

- `tests/test_ingest_mercator_clip.py` monkeypatches
  `_mercator_envelope_degenerates` for the shift-before-a-skipped-clip ordering
  test. `clip_to_mercator_bounds` resolves that name from metadata_mercator's
  globals, so a patch on the façade would have bound a name nothing reads and
  the probe would have run for real.
- `tests/test_crs_degree_agreement.py` imports the two srtext regexes. It is
  the gate on where those predicates live, so it names the file.

`_MODULE_LOC_CAPS`: metadata.py drops 2151 -> 144, exact. No extracted module
is near the 1000-line inclusion threshold, so none needs its own entry yet; the
carve-out history is condensed into the entry's comment. The ruff C901 baseline
entry follows `refresh_attribute_metadata` (the file's only over-complex
function) to metadata_attributes.py — retargeted, not new. The lineage-summary
allowlist in test_analysis_provenance.py follows `_score_metadata_completeness`
to metadata_quality.py.

Cross-references updated on both sides: core/geo.py's two pointers at the
extent writers, ogr.py's four, analysis_sql/shared.py, column_stats.py, and the
four test docstrings that named metadata.py by file.

Closes #1042
