+++ backend/app/processing/ingest/metadata.py
@@ -3,2149 +3,142 @@
 All functions take an AsyncSession and a table name. Table names are validated
 against a strict pattern to prevent SQL injection (they are identifiers, not
 parameterizable values).
+
+fix(#1042): the implementations now live in sibling ``metadata_*`` modules and
+this file re-exports them. It reached 2151 lines with five carve-outs on its
+size cap, each one a correctness fix that had to argue for its lines; the seams
+below are the ones the carve-outs kept landing on.
+
+  - ``metadata_sql``         identifier validation and quoting (the shared base)
+  - ``metadata_geometry``    constructing ``geom``, laundering column names
+  - ``metadata_extent``      reading metadata back off a landed table
+  - ``metadata_mercator``    the clip, the 0..360 shift, the CRS predicates
+  - ``metadata_projection``  the 4326 render column and the reader grant
+  - ``metadata_quality``     dataset quality scoring
+  - ``metadata_attributes``  attribute-metadata rows and their inference
+
+Import from here, not from the sub-modules: this façade is what the ingest
+tasks, the export paths and the extension defaults have always imported, and
+what the ``mock.patch("app.processing.ingest.metadata.<name>")`` targets in the
+test suite resolve against. Two references point at a sub-module on purpose,
+because they need the module a call physically lives in rather than the name it
+is published under: the ``_mercator_envelope_degenerates`` monkeypatch in
+``tests/test_ingest_mercator_clip.py`` and the srtext-predicate imports in
+``tests/test_crs_degree_agreement.py``.
 """
 
-import re
-import uuid
-from datetime import datetime, timezone
-from typing import TYPE_CHECKING
-
-import structlog
-from sqlalchemy import select, text
-from sqlalchemy.exc import DBAPIError
-from sqlalchemy.ext.asyncio import AsyncSession
-
-from app.core.geo import seam_extent_wkt_for_table
-
-if TYPE_CHECKING:
-    from app.core.processing_port import Attribute, Dataset, Record
-    from app.processing.ingest.warnings import MercatorClipCounts
-
-logger = structlog.stdlib.get_logger(__name__)
-
-_TABLE_NAME_RE = re.compile(r"^[a-z0-9_]+$")
-_BOX3D_RE = re.compile(
-    r"^BOX3D\("
-    r"[-+0-9.eE]+\s+[-+0-9.eE]+\s+([-+0-9.eE]+),"
-    r"[-+0-9.eE]+\s+[-+0-9.eE]+\s+([-+0-9.eE]+)"
-    r"\)$"
+from app.processing.ingest.metadata_attributes import (
+    _PG_TYPE_TO_DOMAIN,  # noqa: F401 -- re-exported, see __all__
+    _UNIT_SUFFIX_MAP,  # noqa: F401 -- re-exported, see __all__
+    _build_attribute_metadata,  # noqa: F401 -- re-exported, see __all__
+    _build_geometry_attribute_row,  # noqa: F401 -- re-exported, see __all__
+    _humanize_column_name,  # noqa: F401 -- re-exported, see __all__
+    _infer_domain_type,  # noqa: F401 -- re-exported, see __all__
+    _infer_semantic_role,  # noqa: F401 -- re-exported, see __all__
+    _infer_units,  # noqa: F401 -- re-exported, see __all__
+    generate_attribute_metadata,  # noqa: F401 -- re-exported, see __all__
+    refresh_attribute_metadata,  # noqa: F401 -- re-exported, see __all__
+)
+from app.processing.ingest.metadata_extent import (
+    _ABSTRACT_TO_CONCRETE_GEOMETRY_TYPE,  # noqa: F401 -- re-exported, see __all__
+    _BOX3D_RE,  # noqa: F401 -- re-exported, see __all__
+    _normalize_geometry_type,  # noqa: F401 -- re-exported, see __all__
+    _parse_box3d_z_bounds,  # noqa: F401 -- re-exported, see __all__
+    _seam_crossing_extent_wkt,  # noqa: F401 -- re-exported, see __all__
+    _table_has_geometry,  # noqa: F401 -- re-exported, see __all__
+    detect_3d_metadata,  # noqa: F401 -- re-exported, see __all__
+    extract_metadata,  # noqa: F401 -- re-exported, see __all__
+    get_column_info,  # noqa: F401 -- re-exported, see __all__
+    get_extent,  # noqa: F401 -- re-exported, see __all__
+    get_feature_count,  # noqa: F401 -- re-exported, see __all__
+    get_geometry_type,  # noqa: F401 -- re-exported, see __all__
+    get_sample_values,  # noqa: F401 -- re-exported, see __all__
+    get_table_srid,  # noqa: F401 -- re-exported, see __all__
+    promote_z_to_elev,  # noqa: F401 -- re-exported, see __all__
+)
+from app.processing.ingest.metadata_geometry import (
+    construct_point_geometry,  # noqa: F401 -- re-exported, see __all__
+    construct_wkt_geometry,  # noqa: F401 -- re-exported, see __all__
+    detect_dbf_truncation_collisions,  # noqa: F401 -- re-exported, see __all__
+    ensure_geom_column,  # noqa: F401 -- re-exported, see __all__
+    rename_reserved_columns,  # noqa: F401 -- re-exported, see __all__
+)
+from app.processing.ingest.metadata_mercator import (
+    _DEGREE_UNIT_SRTEXT_RE,  # noqa: F401 -- re-exported, see __all__
+    _GEOGRAPHIC_SRTEXT_RE,  # noqa: F401 -- re-exported, see __all__
+    _MERCATOR_SAFE_ENVELOPE,  # noqa: F401 -- re-exported, see __all__
+    _mercator_envelope_degenerates,  # noqa: F401 -- re-exported, see __all__
+    _shift_zero_to_360_longitudes,  # noqa: F401 -- re-exported, see __all__
+    clip_to_mercator_bounds,  # noqa: F401 -- re-exported, see __all__
+)
+from app.processing.ingest.metadata_projection import (
+    add_4326_column,  # noqa: F401 -- re-exported, see __all__
+    ensure_geom_4326_gist_index,  # noqa: F401 -- re-exported, see __all__
+    grant_reader_access,  # noqa: F401 -- re-exported, see __all__
+    linearize_existing_4326,  # noqa: F401 -- re-exported, see __all__
+)
+from app.processing.ingest.metadata_quality import (
+    _score_attribute_completeness,  # noqa: F401 -- re-exported, see __all__
+    _score_crs,  # noqa: F401 -- re-exported, see __all__
+    _score_geometry_validity,  # noqa: F401 -- re-exported, see __all__
+    _score_metadata_completeness,  # noqa: F401 -- re-exported, see __all__
+    compute_quality_score,  # noqa: F401 -- re-exported, see __all__
+)
+from app.processing.ingest.metadata_sql import (
+    _TABLE_NAME_RE,  # noqa: F401 -- re-exported, see __all__
+    _qtable,  # noqa: F401 -- re-exported, see __all__
+    _sql_quote_ident,  # noqa: F401 -- re-exported, see __all__
+    _validate_table_name,  # noqa: F401 -- re-exported, see __all__
 )
 
-
-def _validate_table_name(table_name: str) -> None:
-    """Validate table name matches safe identifier pattern."""
-    if not _TABLE_NAME_RE.match(table_name):
-        raise ValueError(
-            f"Invalid table name: {table_name!r}. "
-            "Must contain only lowercase letters, digits, and underscores."
-        )
-
-
-def _qtable(table_name: str, schema: str = "data") -> str:
-    """Return quoted '<schema>.table_name' identifier after validation.
-
-    In single_tenant, schema='data' (unchanged behavior).
-    In multi_tenant, callers pass the per-tenant schema from
-    ``tenant_data_schema(current_tenant_var.get())``.
-    Both table_name and schema are validated against the same safe-identifier
-    pattern (lowercase alphanumeric + underscore) before interpolation
-    (T-1209-05: SQL-identifier injection guard).
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)  # schema names follow the same safe pattern
-    return f'"{schema}"."{table_name}"'
-
-
-def _sql_quote_ident(name: str) -> str:
-    """Return a safely double-quoted SQL identifier for use inside text().
-
-    Handles embedded double-quotes by doubling them, which is the
-    PostgreSQL-standard escape. Centralizes the quoting logic that
-    previously lived inline at every call site (PERF-6, KISS).
-
-    fix(#640): colons are backslash-escaped because SQLAlchemy ``text()``
-    parses ``:name`` as a bind parameter even inside double-quoted
-    identifiers (Socrata exports ship columns literally named ``:id``,
-    ``:created_at``, ...). ``text()`` unescapes ``\\:`` back to ``:`` at
-    compile time, so the emitted SQL carries the literal identifier. The
-    output is therefore only valid inside ``text()`` — do not pass it to
-    ``exec_driver_sql`` or raw driver APIs.
-    """
-    return '"' + name.replace('"', '""').replace(":", "\\:") + '"'
-
-
-def _parse_box3d_z_bounds(box3d_text: str | None) -> tuple[float | None, float | None]:
-    """Extract z_min/z_max from a ``BOX3D(...)`` string.
-
-    ``ST_3DExtent`` is widely available in PostGIS stacks where ``ST_Is3D`` may
-    not be. Parsing the aggregate output keeps the ingest pipeline compatible
-    with older extensions while still surfacing useful z-range metadata.
-    """
-    if not box3d_text:
-        return None, None
-
-    match = _BOX3D_RE.match(box3d_text)
-    if match is None:
-        return None, None
-
-    return float(match.group(1)), float(match.group(2))
-
-
-async def construct_point_geometry(
-    session: AsyncSession,
-    table_name: str,
-    x_column: str,
-    y_column: str,
-    srid: int = 4326,
-    *,
-    schema: str = "data",
-) -> int:
-    """Add geometry column from x/y coordinate columns.
-
-    Returns count of rows with valid geometry.
-    """
-    _validate_table_name(table_name)
-    if not _TABLE_NAME_RE.match(x_column) or not _TABLE_NAME_RE.match(y_column):
-        raise ValueError("Invalid column name")
-
-    tref = _qtable(table_name, schema=schema)
-    x_col = _sql_quote_ident(x_column)
-    y_col = _sql_quote_ident(y_column)
-    finite_floor = "-1.7976931348623157e308"
-    finite_ceiling = "1.7976931348623157e308"
-    unparseable = await session.execute(
-        text(
-            f"SELECT COUNT(*) FROM {tref} "
-            f"WHERE {x_col} IS NOT NULL AND {y_col} IS NOT NULL "
-            f"AND (NOT pg_input_is_valid({x_col}::text, 'double precision') "
-            f"OR NOT pg_input_is_valid({y_col}::text, 'double precision'))"
-        )
-    )
-    unparseable_count = int(unparseable.scalar_one())
-    if unparseable_count:
-        raise ValueError(
-            f"{unparseable_count} row(s) contain X/Y values that are not numbers"
-        )
-    if srid == 4326:
-        valid_coordinates = (
-            f"{x_col}::double precision BETWEEN -180 AND 180 AND "
-            f"{y_col}::double precision BETWEEN -90 AND 90"
-        )
-    else:
-        valid_coordinates = (
-            f"{x_col}::double precision BETWEEN {finite_floor} AND {finite_ceiling} "
-            f"AND {y_col}::double precision BETWEEN {finite_floor} AND {finite_ceiling}"
-        )
-    invalid = await session.execute(
-        text(
-            f"SELECT COUNT(*) FROM {tref} "
-            f"WHERE {x_col} IS NOT NULL AND {y_col} IS NOT NULL "
-            f"AND NOT ({valid_coordinates})"
-        )
-    )
-    invalid_count = int(invalid.scalar_one())
-    if invalid_count:
-        coordinate_system = "EPSG:4326 ranges" if srid == 4326 else "finite values"
-        raise ValueError(
-            f"{invalid_count} row(s) contain X/Y coordinates outside {coordinate_system}"
-        )
-    await session.execute(
-        text(f"ALTER TABLE {tref} ADD COLUMN geom geometry(Point, {srid})")
-    )
-    result = await session.execute(
-        text(
-            f"UPDATE {tref} SET geom = ST_SetSRID("
-            f"  ST_MakePoint({x_col}::double precision, {y_col}::double precision), "
-            f"  {srid}) "
-            f"WHERE {x_col} IS NOT NULL AND {y_col} IS NOT NULL "
-            f"AND {valid_coordinates}"
-        )
-    )
-    await session.execute(
-        text(f"CREATE INDEX idx_{table_name}_geom ON {tref} USING GIST (geom)")
-    )
-    # SQLAlchemy CursorResult exposes rowcount for DML; the async Result
-    # type stub is less specific so mypy can't narrow it here.
-    return result.rowcount  # type: ignore[attr-defined]
-
-
-async def construct_wkt_geometry(
-    session: AsyncSession,
-    table_name: str,
-    wkt_column: str,
-    srid: int = 4326,
-    *,
-    schema: str = "data",
-) -> int:
-    """Add geometry column from a WKT text column.
-
-    Returns count of rows with valid geometry.
-    """
-    _validate_table_name(table_name)
-    if not _TABLE_NAME_RE.match(wkt_column):
-        raise ValueError("Invalid column name")
-
-    tref = _qtable(table_name, schema=schema)
-    wkt_col = _sql_quote_ident(wkt_column)
-    parsed_geom = f"ST_GeomFromText({wkt_col}, {srid})"
-    invalid_conditions = [
-        f"lower({wkt_col}) ~ '(nan|inf)'",
-        f"NOT ST_IsValid({parsed_geom})",
-    ]
-    if srid == 4326:
-        invalid_conditions.extend(
-            [
-                f"ST_XMin(Box3D({parsed_geom})) < -180",
-                f"ST_XMax(Box3D({parsed_geom})) > 180",
-                f"ST_YMin(Box3D({parsed_geom})) < -90",
-                f"ST_YMax(Box3D({parsed_geom})) > 90",
-            ]
-        )
-    try:
-        # PostGIS raises on syntactically malformed WKT. A SAVEPOINT keeps that
-        # expected validation failure from aborting the caller's ingest
-        # transaction, allowing it to report a stable ValueError and clean up.
-        async with session.begin_nested():
-            invalid = await session.execute(
-                text(
-                    f"SELECT COUNT(*) FROM {tref} WHERE {wkt_col} IS NOT NULL "
-                    f"AND ({' OR '.join(invalid_conditions)})"
-                )
-            )
-            invalid_count = int(invalid.scalar_one())
-            if invalid_count:
-                coordinate_system = (
-                    "EPSG:4326 ranges" if srid == 4326 else "finite values"
-                )
-                raise ValueError(
-                    f"{invalid_count} row(s) contain invalid WKT geometry or "
-                    f"coordinates outside {coordinate_system}"
-                )
-
-            # Detect geometry type from sample row while malformed input is
-            # still isolated by the validation savepoint.
-            sample = await session.execute(
-                text(
-                    f"SELECT GeometryType({parsed_geom}) "
-                    f"FROM {tref} WHERE {wkt_col} IS NOT NULL LIMIT 1"
-                )
-            )
-            geom_type = sample.scalar_one_or_none() or "GEOMETRY"
-    except DBAPIError as exc:
-        raise ValueError("WKT column contains malformed geometry text") from exc
-
-    await session.execute(
-        text(f"ALTER TABLE {tref} ADD COLUMN geom geometry({geom_type}, {srid})")
-    )
-    result = await session.execute(
-        text(f"UPDATE {tref} SET geom = {parsed_geom} WHERE {wkt_col} IS NOT NULL")
-    )
-    await session.execute(
-        text(f"CREATE INDEX idx_{table_name}_geom ON {tref} USING GIST (geom)")
-    )
-    # SQLAlchemy CursorResult exposes rowcount for DML; the async Result
-    # type stub is less specific so mypy can't narrow it here.
-    return result.rowcount  # type: ignore[attr-defined]
-
-
-async def get_table_srid(
-    session: AsyncSession, table_name: str, schema: str = "data"
-) -> int | None:
-    """Get the SRID of the geom column for a table in the given schema.
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-    result = await session.execute(
-        text("SELECT Find_SRID(:schema, :table_name, 'geom')").bindparams(
-            schema=schema, table_name=table_name
-        )
-    )
-    row = result.scalar_one_or_none()
-    return int(row) if row is not None else None
-
-
-# Phase 1057 WFS-04 layer-2 fix (Phase 1060 close-gate): map abstract OGC
-# GML 3 geometry types (returned by PostGIS GeometryType() when the source
-# WFS stores them, e.g. GeoServer's opengeo:countries) to concrete subtypes
-# that satisfy the chk_datasets_geometry_type CHECK constraint.
-#
-# Background: Phase 1057's ``-nlt GEOMETRY`` fix relaxed the column-type
-# constraint so ogr2ogr could load MultiSurface features without the
-# clip_to_mercator_bounds UPDATE failing. However GeometryType(geom) still
-# returns the actual stored subtype (MULTISURFACE / MULTICURVE / COMPOUND…),
-# and the downstream chk_datasets_geometry_type CHECK constraint only allows
-# the 7 concrete types (POINT, LINESTRING, POLYGON, MULTIPOINT,
-# MULTILINESTRING, MULTIPOLYGON, GEOMETRYCOLLECTION). Until the database
-# layer normalizes the stored geometries themselves, classify the dataset
-# by the closest concrete equivalent of the abstract type — this preserves
-# the user-facing semantics (a country IS a polygon collection) without
-# touching the binary geometry data.
-_ABSTRACT_TO_CONCRETE_GEOMETRY_TYPE: dict[str, str] = {
-    "MULTISURFACE": "MULTIPOLYGON",
-    "MULTICURVE": "MULTILINESTRING",
-    "COMPOUNDCURVE": "MULTILINESTRING",
-    "COMPOUNDSURFACE": "MULTIPOLYGON",
-    "SURFACE": "POLYGON",
-    "CURVE": "LINESTRING",
-    "POLYHEDRALSURFACE": "MULTIPOLYGON",
-    "TIN": "MULTIPOLYGON",
-    "TRIANGLE": "POLYGON",
-}
-
-
-def _normalize_geometry_type(value: str | None) -> str | None:
-    """Normalize abstract OGC geometry type names to concrete subtypes.
-
-    Returns the uppercased input unchanged when it is already a concrete
-    type, ``None`` when the input is ``None`` or empty.
-    """
-    if not value:
-        return None
-    upper = value.upper()
-    return _ABSTRACT_TO_CONCRETE_GEOMETRY_TYPE.get(upper, upper)
-
-
-async def get_geometry_type(
-    session: AsyncSession, table_name: str, *, schema: str = "data"
-) -> str | None:
-    """Get the geometry type of the first feature in the table.
-
-    Returns the type in uppercase for consistent casing across all sources.
-    Abstract GML 3 types (MultiSurface/MultiCurve/etc.) are normalized to the
-    closest concrete equivalent so the value satisfies
-    ``chk_datasets_geometry_type``.
-    """
-    result = await session.execute(
-        text(
-            f"SELECT GeometryType(geom) FROM "
-            f"{_qtable(table_name, schema=schema)} LIMIT 1"
-        )
-    )
-    value = result.scalar_one_or_none()
-    return _normalize_geometry_type(value)
-
-
-async def get_feature_count(
-    session: AsyncSession, table_name: str, *, schema: str = "data"
-) -> int:
-    """Count the number of features (rows) in the table."""
-    result = await session.execute(
-        text(f"SELECT COUNT(*) FROM {_qtable(table_name, schema=schema)}")
-    )
-    return result.scalar_one()
-
-
-async def _seam_crossing_extent_wkt(
-    session: AsyncSession,
-    table_name: str,
-    *,
-    schema: str,
-    xmin: float | None,
-    xmax: float | None,
-) -> str | None:
-    """Two-ring extent WKT when the table honestly crosses ±180, else None.
-
-    fix(#934): gate + delegate for the extent producers below. ``xmin``/``xmax``
-    are the naive ``ST_Extent`` bounds already computed by the caller's query;
-    a crossing dataset folded naively always spans more than 180 degrees (its
-    shifted-domain span is ``360 - width``, so the shifted domain can only win
-    when ``width > 180``), which lets the common non-crossing case skip the
-    second aggregate scan entirely.
-    """
-    if xmin is None or xmax is None:
-        return None
-    if float(xmax) - float(xmin) <= 180.0:
-        return None
-    return await seam_extent_wkt_for_table(session, table_name, schema=schema)
-
-
-async def get_extent(
-    session: AsyncSession, table_name: str, *, schema: str = "data"
-) -> str | None:
-    """Get the 4326 extent WKT (or None for empty tables).
-
-    fix(#934): a Pacific-crossing table (reachable through ordinary ingest
-    since #888 shifts 0..360 sources instead of clipping them) must not store
-    the naive fold — a near-global POLYGON. When the honest extent crosses the
-    seam this returns the two-ring MULTIPOLYGON from ``bbox_to_extent_wkt``;
-    non-crossing extents keep the original query's output byte-identical.
-
-    fix(#430 BA-18): spatial_extent admits POLYGON or MULTIPOLYGON only (fix(#892):
-    typmod geometry(Geometry, 4326) + chk_records_spatial_extent_type). ST_Extent of a
-    single point / axis-collinear points casts to a rejected POINT / LINESTRING, crashing
-    the reupload swap that stores this verbatim. Pad ONLY the degenerate cases into a valid
-    sub-mm polygon; genuine extents stay byte-identical, matching refresh_dataset_metadata.
-    """
-    _validate_table_name(table_name)
-    result = await session.execute(
-        text(
-            f"SELECT CASE "
-            f"  WHEN ext IS NULL THEN NULL "
-            f"  WHEN GeometryType(ext::geometry) = 'POLYGON' "
-            f"    THEN ST_AsText(ST_SetSRID(ext::geometry, 4326)) "
-            f"  ELSE ST_AsText(ST_Expand(ST_SetSRID(ext::geometry, 4326), 1e-9)) "
-            f"END, "
-            f"ST_XMin(ext::geometry), ST_XMax(ext::geometry) "
-            f"FROM (SELECT ST_Extent(geom_4326) AS ext FROM "
-            f"{_qtable(table_name, schema=schema)}) s"
-        )
-    )
-    row = result.one_or_none()
-    if row is None or row[0] is None:
-        return None
-    crossing = await _seam_crossing_extent_wkt(
-        session, table_name, schema=schema, xmin=row[1], xmax=row[2]
-    )
-    return crossing if crossing is not None else row[0]
-
-
-async def detect_3d_metadata(
-    session: AsyncSession, table_name: str, *, schema: str = "data"
-) -> dict:
-    """Detect 3D geometry properties from a PostGIS table.
-
-    Uses ``ST_NDims`` to determine whether any geometry is 3D and
-    ``ST_3DExtent`` to derive z-range metadata.
-    Returns dict with keys: is_3d, n_dims, z_min, z_max.
-    All values are None if the table has no geometry or no rows.
-    """
-    _validate_table_name(table_name)
-
-    _NO_3D = {"is_3d": None, "n_dims": None, "z_min": None, "z_max": None}
-
-    # Check if table has geometry first
-    has_geom = await _table_has_geometry(session, table_name, schema=schema)
-    if not has_geom:
-        return _NO_3D
-
-    try:
-        # Aggregate across all rows to handle mixed-Z datasets correctly
-        result = await session.execute(
-            text(
-                f"SELECT "
-                f"  MAX(ST_NDims(geom)) AS n_dims, "
-                f"  CASE "
-                f"    WHEN MAX(ST_NDims(geom)) > 2 THEN ST_3DExtent(geom)::text "
-                f"    ELSE NULL "
-                f"  END AS extent_3d "
-                f"FROM {_qtable(table_name, schema=schema)} "
-                f"WHERE geom IS NOT NULL"
-            )
-        )
-    except (
-        Exception
-    ):  # broad: ST_NDims/ST_3DExtent may not exist in older PostGIS; degrade gracefully
-        logger.warning(
-            "3d_metadata_detection_failed",
-            table=table_name,
-            hint="ST_NDims or ST_3DExtent may not be available in this PostGIS version",
-        )
-        return _NO_3D
-
-    row = result.one_or_none()
-    if row is None:
-        return {"is_3d": False, "n_dims": 2, "z_min": None, "z_max": None}
-
-    n_dims = row.n_dims if row.n_dims is not None else 2
-    is_3d = bool(n_dims and n_dims > 2)
-    z_min, z_max = _parse_box3d_z_bounds(row.extent_3d if is_3d else None)
-
-    return {
-        "is_3d": is_3d,
-        "n_dims": int(n_dims) if n_dims is not None else None,
-        "z_min": z_min,
-        "z_max": z_max,
-    }
-
-
-async def promote_z_to_elev(
-    session: AsyncSession,
-    table_name: str,
-    geometry_type: str | None,
-    schema: str = "data",
-) -> bool:
-    """For 3D point geometries, extract ST_Z(geom) into an 'elev' numeric column.
-
-    Only runs when:
-    1. The geometry is 3D (caller must verify with detect_3d_metadata first)
-    2. The geometry type is point-like (Point or MultiPoint)
-    3. An 'elev' column does not already exist
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-
-    Returns True if the elev column was created, False otherwise.
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-
-    if geometry_type is None:
-        return False
-
-    # Only promote for point-like geometries
-    geom_upper = geometry_type.upper()
-    if geom_upper not in ("POINT", "MULTIPOINT"):
-        return False
-
-    # Check if elev column already exists
-    col_check = await session.execute(
-        text(
-            "SELECT 1 FROM information_schema.columns "
-            "WHERE table_schema = :schema AND table_name = :t "
-            "AND column_name = 'elev'"
-        ).bindparams(schema=schema, t=table_name)
-    )
-    if col_check.scalar_one_or_none() is not None:
-        return False
-
-    # Add elev column and populate from ST_Z
-    try:
-        await session.execute(
-            text(
-                f"ALTER TABLE {_qtable(table_name, schema=schema)} ADD COLUMN elev double precision"
-            )
-        )
-
-        if geom_upper == "POINT":
-            await session.execute(
-                text(
-                    f"UPDATE {_qtable(table_name, schema=schema)} "
-                    f"SET elev = ST_Z(geom) "
-                    f"WHERE geom IS NOT NULL AND ST_NDims(geom) > 2"
-                )
-            )
-        else:
-            # MultiPoint: extract Z from first point in the multi
-            await session.execute(
-                text(
-                    f"UPDATE {_qtable(table_name, schema=schema)} "
-                    f"SET elev = ST_Z(ST_GeometryN(geom, 1)) "
-                    f"WHERE geom IS NOT NULL AND ST_NDims(geom) > 2"
-                )
-            )
-    except Exception:  # broad: ST_Z/ST_NDims/ST_GeometryN may not be available in older PostGIS; degrade gracefully
-        logger.warning(
-            "promote_z_to_elev_failed",
-            table=table_name,
-            hint="ST_Z, ST_NDims, or ST_GeometryN may not be available",
-        )
-        return False
-
-    return True
-
-
-async def get_column_info(
-    session: AsyncSession, table_name: str, schema: str = "data"
-) -> list[dict]:
-    """Get column names, types, ordinal position, and nullability.
-
-    Excludes internal columns (gid, geom, geom_4326).
-    Returns list of dicts with keys: name, type, ordinal_position, is_nullable.
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-    result = await session.execute(
-        text(
-            "SELECT column_name, data_type, ordinal_position, "
-            "       (is_nullable = 'YES') AS is_nullable "
-            "FROM information_schema.columns "
-            "WHERE table_schema = :schema AND table_name = :table_name "
-            "ORDER BY ordinal_position"
-        ).bindparams(schema=schema, table_name=table_name)
-    )
-    excluded = {"gid", "geom", "geom_4326"}
-    return [
-        {
-            "name": row[0],
-            "type": row[1],
-            "ordinal_position": row[2],
-            "is_nullable": row[3],
-        }
-        for row in result.all()
-        if row[0] not in excluded
-    ]
-
-
-async def get_sample_values(
-    session: AsyncSession,
-    table_name: str,
-    column_info: list[dict],
-    sample_size: int = 10000,
-    schema: str = "data",
-) -> dict:
-    """Extract distinct sample values per column from a data table.
-
-    Returns a dict mapping column name to a list of up to 10 distinct
-    string values. Skips geometry-type columns and columns with no
-    non-null values.
-
-    Implementation: a single CTE pulls ``sample_size`` rows from the base
-    table, then a UNION ALL of branches extracts up to 10 distinct
-    non-null ``::text`` values per column. This is one query and one
-    table scan regardless of column count — replaces the previous N+1
-    per-column query pattern (PERF-1).
-
-    The default ``sample_size`` of 10000 is chosen so that columns which
-    are up to ~99.9% NULL still yield non-empty sample values within the
-    per-column ``LIMIT 10`` display cap. Because the CTE materializes
-    ``sample_size`` rows up-front, base-scan width and peak query RAM
-    grow linearly with this value; bumping it further for even sparser
-    columns should be weighed against the cost on multi-million-row
-    tables. Callers needing narrower sampling can pass ``sample_size``
-    explicitly.
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-
-    # Look up the actual columns in the table so we can filter out any
-    # entries in `column_info` that don't exist (ArcGIS / service ingest
-    # can build column_info from the upstream API schema, which may not
-    # match the landed PostgreSQL table name-for-name — e.g. case
-    # laundering). Missing a single column in the batched query would
-    # error the whole statement, so we must intersect first.
-    live_result = await session.execute(
-        text(
-            "SELECT column_name FROM information_schema.columns "
-            "WHERE table_schema = :schema AND table_name = :t"
-        ).bindparams(schema=schema, t=table_name)
-    )
-    live_columns = {row[0] for row in live_result.all()}
-    if not live_columns:
-        return {}
-
-    # Collect non-geometry columns with their identifier quoted once.
-    candidates: list[tuple[str, str]] = []
-    for col in column_info:
-        col_name = col.get("name", "")
-        col_type = col.get("type", "")
-        if not col_name:
-            continue
-        if "geometry" in col_type.lower():
-            continue
-        if col_name not in live_columns:
-            continue  # silently skip columns that don't exist in the table
-        candidates.append((col_name, _sql_quote_ident(col_name)))
-
-    if not candidates:
-        return {}
-
-    # Build one UNION ALL query that tags each row with its column index.
-    # Column names go into a VALUES lookup table (keyed by index) on the
-    # Python side so the SQL doesn't need to embed arbitrary identifiers
-    # as literals in SELECT aliases.
-    select_cols = ", ".join(q for _, q in candidates)
-    union_branches: list[str] = []
-    for idx, (_, quoted) in enumerate(candidates):
-        union_branches.append(
-            f"(SELECT {idx} AS col_idx, val FROM "
-            f"(SELECT DISTINCT {quoted}::text AS val FROM sampled "
-            f"WHERE {quoted} IS NOT NULL LIMIT 10) s)"
-        )
-    union_sql = " UNION ALL ".join(union_branches)
-    # Use TABLESAMPLE BERNOULLI for representative sampling on large tables.
-    # For small tables (<500 rows), BERNOULLI(10) may return 0 rows, so we
-    # fall back to a plain LIMIT which is fine for small datasets.
-    tbl = _qtable(table_name, schema=schema)
-    query = (
-        f"WITH sampled AS ("
-        f"  SELECT {select_cols} FROM {tbl}"
-        f"  TABLESAMPLE BERNOULLI ("
-        f"    CASE WHEN (SELECT c.reltuples FROM pg_class c "
-        f"      JOIN pg_namespace n ON n.oid = c.relnamespace "
-        f"      WHERE c.relname = :t AND n.nspname = :schema) > 500"
-        f"         THEN 10 ELSE 100 END"
-        f"  ) LIMIT :sample_size"
-        f") "
-        f"{union_sql}"
-    )
-
-    rows = await session.execute(
-        text(query).bindparams(sample_size=sample_size, t=table_name, schema=schema)
-    )
-
-    result: dict[str, list[str]] = {}
-    for row in rows.all():
-        idx, val = row[0], row[1]
-        if val is None:
-            continue
-        col_name = candidates[idx][0]
-        result.setdefault(col_name, []).append(val)
-
-    return result
-
-
-async def _score_metadata_completeness(
-    session: AsyncSession,
-    record: "Record",
-) -> float:
-    """Percentage of optional metadata fields that are populated (0-100)."""
-    from app.platform.extensions import get_processing_port
-
-    port = get_processing_port()
-    kw_count = await port.get_record_keyword_count(session, record.id)
-    has_keywords = True if kw_count and kw_count > 0 else None
-
-    optional_fields = [
-        record.summary,
-        has_keywords,
-        record.license,
-        record.source_organization,
-        record.temporal_start,
-        record.lineage_summary,
-        record.update_frequency,
-        record.usage_constraints,
-        record.access_constraints,
-        record.theme_category if record.theme_category else None,
-    ]
-    filled = sum(1 for f in optional_fields if f is not None)
-    return round(filled / len(optional_fields) * 100, 1)
-
-
-def _score_crs(dataset: "Dataset") -> float:
-    """100 if SRID is defined or dataset has no geometry, else 0."""
-    has_geometry = dataset.geometry_type is not None
-    return 100.0 if (dataset.srid is not None or not has_geometry) else 0.0
-
-
-async def _score_geometry_validity(
-    session: AsyncSession,
-    table_name: str,
-    has_geometry: bool,
-    max_rows: int,
-    *,
-    schema: str = "data",
-) -> float:
-    """Percentage of valid geometries (0-100). Degrades to 100 on error."""
-    if not has_geometry:
-        return 100.0
-    try:
-        async with session.begin_nested():
-            result = await session.execute(
-                text(
-                    f"SELECT COUNT(*) FILTER (WHERE ST_IsValid(geom)) * 100.0 / NULLIF(COUNT(*), 0) "
-                    f"FROM (SELECT geom FROM "
-                    f"{_qtable(table_name, schema=schema)} LIMIT :max_rows) sub"
-                ).bindparams(max_rows=max_rows)
-            )
-            val = result.scalar_one_or_none()
-            if val is not None:
-                return round(float(val), 1)
-    except (
-        Exception
-    ):  # broad: ST_IsValid quality score is non-fatal; degrade to 100.0 on any DB error
-        pass
-    return 100.0
-
-
-async def _score_attribute_completeness(
-    session: AsyncSession,
-    table_name: str,
-    column_info: list[dict],
-    *,
-    schema: str = "data",
-) -> float:
-    """Average non-null percentage across non-geometry columns (0-100)."""
-    non_geom_cols = [
-        c
-        for c in column_info
-        if "geometry" not in c.get("type", "").lower() and c.get("name")
-    ]
-    if not non_geom_cols:
-        return 100.0
-    col_exprs = ", ".join(
-        f"COUNT({_sql_quote_ident(col['name'])}) "
-        f'* 100.0 / NULLIF(COUNT(*), 0) AS "s_{i}"'
-        for i, col in enumerate(non_geom_cols)
-    )
-    try:
-        async with session.begin_nested():
-            result = await session.execute(
-                text(f"SELECT {col_exprs} FROM {_qtable(table_name, schema=schema)}")
-            )
-            row = result.one_or_none()
-            if row is not None:
-                col_scores: list[float] = [float(v) for v in row if v is not None]
-                if col_scores:
-                    return round(sum(col_scores) / len(col_scores), 1)
-    except Exception:  # broad: attribute completeness score is non-fatal; degrade to 100.0 on any DB error
-        pass
-    return 100.0
-
-
-async def compute_quality_score(
-    session: AsyncSession,
-    table_name: str,
-    column_info: list[dict],
-    dataset: "Dataset",
-    max_validity_rows: int = 10000,
-    *,
-    schema: str = "data",
-) -> dict:
-    """Compute a weighted quality score for a dataset.
-
-    Dimensions:
-    - Metadata completeness (30%): non-empty optional fields on dataset
-    - Geometry validity (30%): percentage of valid geometries
-    - Attribute completeness (25%): average non-null percentage across columns
-    - CRS defined (15%): 100 if srid is set, else 0
-
-    Returns a dict with overall score and per-dimension scores.
-    """
-    _validate_table_name(table_name)
-    record = dataset.record
-    has_geometry = dataset.geometry_type is not None
-
-    metadata_score = await _score_metadata_completeness(session, record)
-    crs_score = _score_crs(dataset)
-    geometry_score = await _score_geometry_validity(
-        session,
-        table_name,
-        has_geometry,
-        max_validity_rows,
-        schema=schema,
-    )
-    attribute_score = await _score_attribute_completeness(
-        session,
-        table_name,
-        column_info,
-        schema=schema,
-    )
-
-    # For table records, geometry_validity and crs_defined are not applicable.
-    # Re-normalize weights: metadata (30) + attribute (25) = 55 total.
-    is_table = getattr(record, "record_type", None) == "table"
-    if is_table:
-        overall = round(metadata_score * (30 / 55) + attribute_score * (25 / 55))
-        return {
-            "overall": overall,
-            "metadata_completeness": metadata_score,
-            "attribute_completeness": attribute_score,
-            "geometry_validity": None,
-            "crs_defined": None,
-            "computed_at": datetime.now(timezone.utc).isoformat(),
-        }
-
-    overall = round(
-        metadata_score * 0.30
-        + geometry_score * 0.30
-        + attribute_score * 0.25
-        + crs_score * 0.15
-    )
-
-    return {
-        "overall": overall,
-        "metadata_completeness": metadata_score,
-        "geometry_validity": geometry_score,
-        "attribute_completeness": attribute_score,
-        "crs_defined": crs_score,
-        "computed_at": datetime.now(timezone.utc).isoformat(),
-    }
-
-
-async def _table_has_geometry(
-    session: AsyncSession, table_name: str, schema: str = "data"
-) -> bool:
-    """Check whether a table has a 'geom' column in the given schema.
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-    result = await session.execute(
-        text(
-            "SELECT EXISTS(SELECT 1 FROM information_schema.columns "
-            "WHERE table_schema=:schema AND table_name=:table_name "
-            "AND column_name='geom')"
-        ).bindparams(schema=schema, table_name=table_name)
-    )
-    return result.scalar_one()
-
-
-async def extract_metadata(
-    session: AsyncSession, table_name: str, schema: str = "data"
-) -> dict:
-    """Extract all metadata from a PostGIS table.
-
-    PERF-03 (Phase 274): for spatial tables, the four data-table SELECTs
-    (feature_count, srid, geometry_type, extent_wkt) are consolidated
-    into a single CTE so the database does one shared scan and one
-    round-trip. For non-spatial tables only feature_count is queried.
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-
-    Returns dict with keys: srid, geometry_type, feature_count, extent_wkt,
-    column_info. For non-spatial tables, spatial fields are None.
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-    column_info = await get_column_info(session, table_name, schema=schema)
-    has_geometry = await _table_has_geometry(session, table_name, schema=schema)
-
-    if not has_geometry:
-        feature_count = await get_feature_count(session, table_name, schema=schema)
-        return {
-            "srid": None,
-            "geometry_type": None,
-            "feature_count": feature_count,
-            "extent_wkt": None,
-            "column_info": column_info,
-        }
-
-    # Spatial-table fast path: single CTE pulls feature_count, srid,
-    # geometry_type, and extent_wkt in one query. Mirrors the original
-    # helpers' semantics:
-    #   - feature_count: SELECT COUNT(*) FROM {schema}.<t>
-    #   - srid: Find_SRID(:schema, :t, 'geom')
-    #   - geometry_type: GeometryType(geom) FROM {schema}.<t> LIMIT 1
-    #     (NOTE: original returns None when zero rows; we mirror that.)
-    #   - extent_wkt: ST_AsText(ST_SetSRID(ST_Extent(geom_4326)::geometry, 4326))
-    #     (uppercased for None when no rows.)
-    tref = _qtable(table_name, schema=schema)
-    try:
-        result = await session.execute(
-            text(
-                f"""
-                WITH meta AS (
-                    SELECT
-                        COUNT(*) AS feature_count,
-                        (
-                            SELECT GeometryType(geom)
-                            FROM {tref}
-                            WHERE geom IS NOT NULL
-                            LIMIT 1
-                        ) AS geometry_type,
-                        -- fix(#430 BA-18): pad only degenerate (point/line) extents
-                        -- into a POLYGON; spatial_extent rejects POINT/LINESTRING.
-                        CASE
-                            WHEN ST_Extent(geom_4326) IS NULL THEN NULL
-                            WHEN GeometryType(ST_Extent(geom_4326)::geometry) = 'POLYGON'
-                            THEN ST_AsText(
-                                ST_SetSRID(ST_Extent(geom_4326)::geometry, 4326)
-                            )
-                            ELSE ST_AsText(
-                                ST_Expand(
-                                    ST_SetSRID(ST_Extent(geom_4326)::geometry, 4326),
-                                    1e-9
-                                )
-                            )
-                        END AS extent_wkt,
-                        ST_XMin(ST_Extent(geom_4326)) AS ext_xmin,
-                        ST_XMax(ST_Extent(geom_4326)) AS ext_xmax
-                    FROM {tref}
-                )
-                SELECT
-                    feature_count,
-                    geometry_type,
-                    extent_wkt,
-                    ext_xmin,
-                    ext_xmax,
-                    Find_SRID(:schema, :t, 'geom') AS srid
-                FROM meta
-                """
-            ).bindparams(schema=schema, t=table_name)
-        )
-        row = result.one()
-        # Phase 1057 WFS-04 layer-2 fix: normalize abstract OGC GML 3 types.
-        geometry_type = _normalize_geometry_type(row.geometry_type)
-        # fix(#934): a Pacific-crossing source must not store the naive fold
-        # (a near-global POLYGON); emit the two-ring MULTIPOLYGON instead.
-        # Non-crossing extents keep the CTE's output byte-identical.
-        extent_wkt = row.extent_wkt
-        crossing = await _seam_crossing_extent_wkt(
-            session, table_name, schema=schema, xmin=row.ext_xmin, xmax=row.ext_xmax
-        )
-        if crossing is not None:
-            extent_wkt = crossing
-        return {
-            "srid": int(row.srid) if row.srid is not None else None,
-            "geometry_type": geometry_type,
-            "feature_count": row.feature_count,
-            "extent_wkt": extent_wkt,
-            "column_info": column_info,
-        }
-    except Exception:  # broad: degrade to the per-helper path on any DB-level error
-        # PERF-03 fallback: some PostGIS deployments lack Find_SRID, or
-        # ST_Extent may fail with a malformed geometry. Mirror the
-        # original ordering of the four helpers so behavior is identical.
-        logger.warning(
-            "extract_metadata_cte_failed_falling_back",
-            table=table_name,
-            exc_info=True,
-        )
-        srid = await get_table_srid(session, table_name, schema=schema)
-        geometry_type = await get_geometry_type(session, table_name, schema=schema)
-        extent_wkt = await get_extent(session, table_name, schema=schema)
-        feature_count = await get_feature_count(session, table_name, schema=schema)
-        return {
-            "srid": srid,
-            "geometry_type": geometry_type,
-            "feature_count": feature_count,
-            "extent_wkt": extent_wkt,
-            "column_info": column_info,
-        }
-
-
-async def ensure_geom_column(
-    session: AsyncSession, table_name: str, schema: str = "data"
-) -> bool:
-    """Rename the geometry column to 'geom' if ogr2ogr used a different name.
-
-    In the happy path this renames the `_geolens_geom` placeholder that
-    `run_ogr2ogr` / `run_ogr2ogr_service` create (see the GEOMETRY_NAME
-    override in ogr.py) to `geom`. It also handles legacy edge cases where
-    ogr2ogr creates 'wkb_geometry' instead (e.g. when appending to a
-    pre-existing table or when a driver ignores -lco GEOMETRY_NAME).
-
-    Must run AFTER `rename_reserved_columns` so that any source attribute
-    named `geom`/`geometry` has already been moved to `src_<name>`,
-    leaving `geom` free for the rename.
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-
-    Returns True if the table has a geometry column, False for non-spatial tables.
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-    result = await session.execute(
-        text(
-            "SELECT f_geometry_column FROM geometry_columns "
-            "WHERE f_table_schema = :schema AND f_table_name = :table_name"
-        ),
-        {"schema": schema, "table_name": table_name},
-    )
-    row = result.first()
-    if row is None:
-        return False  # Non-spatial table
-
-    geom_col = row[0]
-    if geom_col == "geom":
-        return True  # Already correct
-
-    logger.info(
-        "Renaming geometry column",
-        table=table_name,
-        from_col=geom_col,
-        to_col="geom",
-    )
-    _validate_table_name(geom_col)
-    await session.execute(
-        text(
-            f"ALTER TABLE {_qtable(table_name, schema=schema)} "
-            f"RENAME COLUMN {_sql_quote_ident(geom_col)} TO geom"
-        )
-    )
-    # ING-02 / P2-02 (Phase 1076): no internal commit. The caller
-    # (_finalize_ingest at tasks_common.py:821) owns the phase-2 commit
-    # boundary so a downstream failure rolls back this rename atomically.
-    return True
-
-
-async def rename_reserved_columns(
-    session: AsyncSession,
-    table_name: str,
-    schema: str = "data",
-) -> list[dict]:
-    """Rename any source column whose name collides with a GeoLens-internal
-    PostGIS column (gid, geom, geometry, geom_4326, fid, ogc_fid) to
-    ``src_<name>``. Runs BEFORE add_4326_column so that ALTER TABLE ADD COLUMN
-    geom_4326 cannot collide with a source attribute.
-
-    fix(#640): also renames columns containing ``:`` (Socrata exports ship
-    system columns literally named ``:id``, ``:created_at``, ... which
-    survive OGR's laundering). A colon inside a double-quoted identifier is
-    parsed as a bind parameter by SQLAlchemy ``text()``, so such names break
-    every downstream text()-built query (sampling, column stats, tiles).
-    They are laundered to letter-leading safe names (``:id`` -> ``id``).
-
-    Only renames columns that were NOT created by the ingest pipeline itself:
-    - ``gid``: pipeline creates it as a serial PRIMARY KEY (column_default is
-      non-null). A source-origin ``gid`` has no default and is not an identity.
-    - ``geom`` / ``geometry``: pipeline creates a PostGIS geometry column
-      (data_type = 'USER-DEFINED', udt_name = 'geometry'). Any other type is
-      source-origin.
-    - ``geom_4326``: always renamed on entry — this helper runs before
-      add_4326_column, so any existing ``geom_4326`` must be source-origin.
-    - ``fid``, ``ogc_fid``: always renamed (ogr2ogr with -lco FID=gid does not
-      create these; any such column is source-origin).
-
-    Returns a list of rename records ``[{"original": "gid", "renamed": "src_gid"}, ...]``
-    which callers can attach to ``job.user_metadata['warnings']``.
-    """
-    from app.processing.ingest.ogr import RESERVED_COLUMN_NAMES
-
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-
-    # PERF-4: fast-path — most ingests have zero reserved-name collisions,
-    # so check first with a tiny WHERE-filtered query before fetching the
-    # full column list. Skips the full-table scan in the common case.
-    reserved_check = await session.execute(
-        text(
-            "SELECT column_name "
-            "FROM information_schema.columns "
-            "WHERE table_schema = :schema AND table_name = :t "
-            "AND (column_name = ANY(:names) OR strpos(column_name, ':') > 0)"
-        ).bindparams(schema=schema, t=table_name, names=list(RESERVED_COLUMN_NAMES))
-    )
-    if not reserved_check.first():
-        return []
-
-    # At least one collision candidate exists — now fetch everything we
-    # need to decide whether each candidate is source-origin and what
-    # rename target is safe.
-    result = await session.execute(
-        text(
-            "SELECT column_name, data_type, udt_name, column_default, is_identity "
-            "FROM information_schema.columns "
-            "WHERE table_schema = :schema AND table_name = :t "
-            "ORDER BY ordinal_position"
-        ).bindparams(schema=schema, t=table_name)
-    )
-    rows = result.all()
-    all_column_names = {r[0] for r in rows}
-
-    # Wrap the entire rename loop in a SAVEPOINT so a mid-loop ALTER
-    # failure rolls back ALL previously-applied renames atomically (R-4).
-    # Without this, an ALTER failing on the Nth rename would commit N-1
-    # renames and leave the ingest table in an inconsistent state that
-    # the caller cannot recover from.
-    renames: list[dict] = []
-    try:
-        async with session.begin_nested():
-            for row in rows:
-                col_name, data_type, udt_name, col_default, is_identity = row
-                if col_name not in RESERVED_COLUMN_NAMES and ":" not in col_name:
-                    continue
-
-                if ":" in col_name:
-                    # fix(#640): colon-bearing (Socrata-style) column —
-                    # launder to a safe name. Must start with a letter or the
-                    # column-stats/distinct-values identifier validator
-                    # rejects it (codex P2 on #646): ":id" -> "id", not "_id".
-                    base = re.sub(r"[^A-Za-z0-9_]", "_", col_name).strip("_")
-                    if not base or not base[0].isalpha():
-                        base = f"col_{base}" if base else "col"
-                    # A laundered name may hit an internal name (":geom" ->
-                    # "geom") — apply the reserved-name rule so the staging
-                    # pipeline's own geometry columns stay uncontested
-                    # (codex P2 round 2 on #646).
-                    if base in RESERVED_COLUMN_NAMES:
-                        base = f"src_{base}"
-                    base = base[:63]
-                else:
-                    # Determine if this column was created by the pipeline or came from the source.
-                    if col_name == "gid":
-                        # Pipeline-created gid is a serial/identity with a nextval default.
-                        # Source-origin gid has no default and is not an identity column.
-                        is_pipeline_gid = (
-                            col_default is not None and "nextval" in str(col_default)
-                        ) or (is_identity == "YES")
-                        if is_pipeline_gid:
-                            continue  # This is the pipeline's own gid — leave it alone.
-
-                    elif col_name in ("geom", "geometry"):
-                        # Pipeline-created geometry column has data_type = 'USER-DEFINED'
-                        # and udt_name = 'geometry'. Source-origin columns have other types.
-                        if data_type == "USER-DEFINED" and udt_name == "geometry":
-                            continue  # Pipeline-created spatial column — leave it alone.
-
-                    # All remaining reserved-name columns are source-origin. Rename to src_<name>.
-                    base = f"src_{col_name}"
-
-                # If the target already exists, append a numeric suffix.
-                target = base
-                suffix = 2
-                while target in all_column_names:
-                    target = f"{base[:60]}_{suffix}"
-                    suffix += 1
-
-                # Execute the rename using double-quoted identifiers (not bindable).
-                q_orig = _sql_quote_ident(col_name)
-                q_target = _sql_quote_ident(target)
-                await session.execute(
-                    text(
-                        f'ALTER TABLE "{schema}"."{table_name}" '
-                        f"RENAME COLUMN {q_orig} TO {q_target}"
-                    )
-                )
-
-                logger.warning(
-                    "Renamed reserved or unsafe source column",
-                    table=table_name,
-                    original=col_name,
-                    renamed=target,
-                )
-                renames.append({"original": col_name, "renamed": target})
-                # Update the in-memory set so subsequent iterations see the new name.
-                all_column_names.discard(col_name)
-                all_column_names.add(target)
-    except Exception as exc:  # broad: ALTER TABLE can fail for schema/permission reasons; re-raise to fail the job
-        # Savepoint rollback already unwound any partial renames; re-raise so
-        # the caller's exception handler marks the ingest job as failed with
-        # a clear error message.
-        logger.error(
-            "Reserved-column rename failed; table left in pre-rename state",
-            table=table_name,
-            error=str(exc),
-            renames_attempted=len(renames),
-            exc_info=True,
-        )
-        raise
-
-    if renames:
-        await session.commit()
-    return renames
-
-
-def detect_dbf_truncation_collisions(
-    source_columns: list[dict],
-) -> list[dict]:
-    """Detect shapefile DBF 10-character field-name truncation collisions.
-
-    Given the source-file column list from run_ogrinfo_preview(), returns a
-    list of collision records grouped by the first 10 lowercase characters:
-      [{"truncated": "population", "originals": ["population_2020", "population_2021"]}]
-
-    Only returns groups with 2+ original names — a single column is not a
-    collision. Empty input returns an empty list.
-    """
-    truncation_map: dict[str, list[str]] = {}
-    for col in source_columns:
-        name = col.get("name", "")
-        truncated = name[:10].lower()
-        truncation_map.setdefault(truncated, []).append(name)
-
-    return [
-        {"truncated": truncated, "originals": originals}
-        for truncated, originals in truncation_map.items()
-        if len(originals) >= 2
-    ]
-
-
-# Web Mercator (EPSG:3857) cannot represent latitudes beyond ±85.06°.
-# Geometries extending past this (e.g. Antarctica at -90°) cause
-# "transform: tolerance condition error" in ST_Transform.
-#
-# fix(#899 codex r1): this is a box, not a latitude cutoff — it bounds X at
-# ±180 as well. A point at lon 400 is dropped by the X bound with a perfectly
-# ordinary latitude, so the warning built from these counts must not tell the
-# user that latitude was the problem.
-_MERCATOR_SAFE_ENVELOPE = "ST_MakeEnvelope(-180, -85.06, 180, 85.06, 4326)"
-
-# fix(#888): matches the WKT1 (GEOGCS) and WKT2 (GEOGCRS) spellings PostGIS
-# ships in spatial_ref_sys.srtext for lon/lat CRSs. 4326/4979/4269 match;
-# 2263/3857 (projected, X in feet/metres) do not.
-#
-# fix(#961 review): the ANCHOR is load-bearing and must stay, even though it
-# makes this predicate disagree with `core.geo.wkt_is_geographic` and with the
-# degenerate-envelope floor below on wrapped CRSs (BOUNDCRS, COMPD_CS). Those
-# two can afford to see through a wrapper; this one cannot, and the reason is
-# the same one `core.geo._parse_crs` documents at length: a flat scan cannot
-# decide which subtree a token belongs to.
-#
-# Concretely. Un-anchoring lets `BOUNDCRS[SOURCECRS[GEOGCS[... UNIT["grad"...`
-# match as geographic, and `_DEGREE_UNIT_SRTEXT_RE` below is a flat substring
-# scan that would then find an unrelated `degree` on the TARGET CRS or a
-# PRIMEM and report degrees. `_shift_zero_to_360_longitudes` would subtract
-# 360 from coordinates whose full turn is 400 grads — silent corruption, and
-# the shape is real enough that `tests/test_wkt_is_geographic.py::
-# test_boundcrs_reports_the_source_crs_units_not_the_targets` already pins it
-# on the Python side.
-#
-# Declining to shift is safe: the source is then clipped and REPORTED by
-# `clip_to_mercator_bounds`'s accounting, which is a visible outcome. Shifting
-# wrongly is not. So this gate stays deliberately incomplete, and the property
-# tests/test_crs_degree_agreement.py enforces is soundness — it never fires
-# where PROJ says the axes are not degrees — rather than agreement.
-_GEOGRAPHIC_SRTEXT_RE = "^GEOG(CS|CRS)"
-
-# fix(#899 codex r1): geographic is not the same as degree-based. 14 SRIDs in a
-# stock PostGIS spatial_ref_sys are GEOGCS with an angular unit of grads — the
-# Paris-meridian family, 4807 NTF (Paris) and relatives — where a full circle
-# is 400, not 360. Translating one of those by -360 would move a valid feature
-# to a wrong place, so the unit has to be degrees before anything shifts. The
-# pattern covers WKT1 (`UNIT["degree"`) and WKT2 (`ANGLEUNIT["degree"`) in one
-# go, since the WKT2 spelling contains the WKT1 substring. The prefix test
-# above is what keeps a projected CRS out: 3857's srtext also carries
-# `UNIT["degree"` inside its nested GEOGCS.
-_DEGREE_UNIT_SRTEXT_RE = 'UNIT\\["degree'
-
-
-async def _shift_zero_to_360_longitudes(
-    session: AsyncSession, table_name: str, schema: str, src_srid: int
-) -> bool:
-    """Shift a 0..360-convention source into -180..180. True when it shifted.
-
-    fix(#888): a source written in the 0..360 Pacific convention (common in
-    ocean and climate data) is not out-of-range data — it is the same world
-    with a different origin. Clipping it to the Mercator envelope silently
-    deletes everything east of lon 180; translating it preserves every
-    feature. #883 showed that a single-condition guard on this class of
-    problem is a coin flip, so *all four* of these must hold before anything
-    moves:
-
-    1. The geometry column's CRS is lon/lat AND its angular unit is degrees.
-       "Longitude" is meaningless in a projected CRS, where X is metres or feet
-       and 300 is not out of range; and in a grads-based geographic CRS a full
-       circle is 400, so -360 is not a whole turn (fix(#899 codex r1)).
-    2. Table-wide min X >= 0. Any negative longitude means the source is
-       already -180..180; a source mixing both conventions is ambiguous, so
-       refuse to guess.
-    3. Table-wide max X > 180. This is the condition that separates a real
-       0..360 source from a dataset legitimately confined to the eastern
-       hemisphere (0..180 — Africa/Europe/Asia), which must not be flung
-       into -360..-180.
-    4. Table-wide max X <= 360. Past 360 is not the 0..360 convention at all
-       (wrong units, corrupt coordinates); leave those to the clamp.
-
-    Only rows whose *own* min X is >= 180 are translated, so a feature that
-    is already inside -180..180 keeps its exact coordinates. A feature that
-    straddles lon 180 in such a source needs an antimeridian split (#884 /
-    #886) rather than a translate: it stays put and is then reported by the
-    clip accounting in ``clip_to_mercator_bounds``.
-    """
-    tref = _qtable(table_name, schema=schema)
-
-    is_degree_lonlat = await session.scalar(
-        text(
-            "SELECT srtext ~* :geographic AND srtext ~* :degree_unit "
-            "FROM spatial_ref_sys WHERE srid = :srid"
-        ).bindparams(
-            geographic=_GEOGRAPHIC_SRTEXT_RE,
-            degree_unit=_DEGREE_UNIT_SRTEXT_RE,
-            srid=src_srid,
-        )
-    )
-    if not is_degree_lonlat:
-        return False
-
-    # The raw (unfolded) coordinate range is exactly what is wanted here —
-    # this is a convention probe, not a geographic extent, so it is not the
-    # antimeridian-naive extent fold tracked by #886.
-    bounds = (
-        await session.execute(
-            text(
-                f"SELECT ST_XMin(bb), ST_XMax(bb) FROM (SELECT ST_Extent(geom) AS bb FROM {tref}) s"
-            )
-        )
-    ).first()
-    if bounds is None or bounds[0] is None or bounds[1] is None:
-        return False
-    min_x, max_x = float(bounds[0]), float(bounds[1])
-    if min_x < 0 or max_x <= 180 or max_x > 360:
-        return False
-
-    result = await session.execute(
-        text(
-            f"UPDATE {tref} SET geom = ST_Translate(geom, -360, 0) "
-            f"WHERE geom IS NOT NULL AND NOT ST_IsEmpty(geom) AND ST_XMin(geom) >= 180"
-        )
-    )
-    logger.info(
-        "Shifted 0..360 longitudes into -180..180 before the Mercator clip",
-        table=table_name,
-        schema=schema,
-        srid=src_srid,
-        min_x=min_x,
-        max_x=max_x,
-        rows_shifted=result.rowcount,
-    )
-    return True
-
-
-async def _mercator_envelope_degenerates(
-    session: AsyncSession, table_name: str, schema: str, src_srid: int
-) -> bool:
-    """True when the safe envelope collapses under ST_Transform into ``src_srid``.
-
-    fix(#906): for a CRS with a narrow area of validity, transforming the
-    global Mercator safe envelope collapses it — in EPSG:4807 (NTF Paris,
-    grads) every X of the envelope becomes 197.396 — and the clip's
-    intersection would then silently empty the entire table. Enumerated
-    against a stock ``spatial_ref_sys``: 4415 of 8500 SRIDs produce a
-    zero-area or collapsed envelope and another 108 error outright, so this
-    is a class, not an exotic corner.
-
-    Degenerate means any of:
-
-    - zero (or negative) area, or a collapsed X/Y range — the hard collapse
-      (EPSG:4807, every UTM zone, France's 2154, most national grids);
-    - a sliver: area under 1e-6 of its own bbox area — EPSG:2263 leaves a
-      19-square-foot bowtie stretched across a 3e16 sq ft bbox, and EPSG:5070
-      a 0.04 m² one; positive area, still data-destroying;
-    - for a projected CRS, an envelope bbox under 1000 linear units in either
-      dimension — EPSG:27700 collapses to a 0.005 m² *square* (ratio 1, so
-      the sliver test misses it) and polar stereographic 3031 to ~4 m across.
-      The floor is absolute, not relative to the table's extent, because a
-      3857 table with features genuinely beyond ±20 037 508 m is WIDER than
-      its correctly transformed envelope and is exactly what the clip exists
-      to trim (fix(#906 codex r1)). 1000 is orders of magnitude under any
-      sane world envelope in any projected linear unit (metres, feet, even
-      kilometres give ~40 075) and orders over every measured collapse.
-      Guarded to projected CRSs (see the four-mechanism table on #939)
-      because geographic units are degrees (the world is 360 wide) and a
-      geographic transform cannot collapse. The geographic test sees
-      through COMPD_CS (fix(#906 codex r2)): a compound geographic CRS like
-      stock 5498 (NAD83 + NAVD88) is degrees despite its prefix;
-    - the transform raising — the clip's own UPDATE would raise identically,
-      turning a bounds check into a failed ingest. Probed inside a SAVEPOINT
-      so the surrounding phase-2 transaction stays usable.
-    """
-    probe = text(
-        f"WITH env AS (SELECT ST_Transform({_MERCATOR_SAFE_ENVELOPE}, :srid) AS e) "
-        f"SELECT "
-        f"  e IS NULL OR ST_IsEmpty(e) OR ST_Area(e) <= 0 "
-        f"  OR ST_XMin(e) >= ST_XMax(e) OR ST_YMin(e) >= ST_YMax(e) "
-        f"  OR ST_Area(e) < 1e-6 * ((ST_XMax(e) - ST_XMin(e)) "
-        f"                        * (ST_YMax(e) - ST_YMin(e))) "
-        f"  OR ( "
-        # fix(#906 codex r2): the floor's geographic test must see through
-        # COMPD_CS — a compound geographic CRS (e.g. stock 5498, NAD83 +
-        # NAVD88) starts with COMPD_CS but its horizontal axes are degrees,
-        # and its valid 360x170-degree envelope would trip the 1000-unit
-        # floor. Geographic-horizontal here means: a GEOG keyword present and
-        # no PROJ keyword anywhere (every projected WKT1 nests a GEOGCS, so
-        # the PROJ test must win) — the same keyword logic as
-        # core.geo.wkt_is_geographic, in srtext form.
-        # fix(#961 review): this predicate and the 0..360 gate's
-        # `_GEOGRAPHIC_SRTEXT_RE` deliberately DISAGREE on wrapped CRSs, and
-        # unifying them was tried and reverted. Seeing through a wrapper is
-        # right here (the consequence is a size floor) and unsafe there (the
-        # consequence is translating geometry by 360 in a CRS whose turn may
-        # be 400 grads). The asymmetry is the decision, not an oversight; both
-        # halves are pinned by tests/test_crs_degree_agreement.py.
-        f"    NOT COALESCE((SELECT srtext ~* 'GEOG(CS|CRS)' "
-        f"                     AND srtext !~* 'PROJ(CS|CRS)' "
-        f"                  FROM spatial_ref_sys "
-        f"                  WHERE srid = :srid), false) "
-        f"    AND LEAST(ST_XMax(e) - ST_XMin(e), ST_YMax(e) - ST_YMin(e)) < 1000) "
-        f"FROM env"
-    ).bindparams(srid=src_srid)
-    try:
-        async with session.begin_nested():
-            result = await session.execute(probe)
-            return bool(result.scalar_one())
-    except DBAPIError:
-        logger.warning(
-            "Mercator envelope transform failed while probing for degeneracy; "
-            "treating the envelope as unusable in this CRS",
-            table=table_name,
-            schema=schema,
-            srid=src_srid,
-            exc_info=True,
-        )
-        return True
-
-
-async def clip_to_mercator_bounds(
-    session: AsyncSession, table_name: str, schema: str = "data"
-) -> "MercatorClipCounts | None":
-    """Clip geometries to the Web Mercator safe envelope (±85.06° lat).
-
-    Only updates rows whose geometry actually extends beyond the bounds,
-    so this is a no-op for most datasets.
-
-    ``schema`` defaults to ``"data"`` for single_tenant backward compatibility.
-    In multi_tenant callers pass ``_current_tenant_schema()`` (CR-03, Phase 1209).
-
-    Two CRS-related quirks the SQL has to handle:
-
-    1. Envelope is in SRID 4326. If the column's SRID differs (4979 / any
-       projected CRS), transform the envelope to match — otherwise PostGIS
-       raises `coveredby: Operation on mixed SRID geometries`.
-    2. The envelope is always 2D. If the column is declared 3D (e.g.
-       `MultiPointZ` for a 4979 source with elevation), `ST_Intersection`
-       drops Z and the UPDATE then fails with `Column has Z dimension but
-       geometry does not`. Wrap the result in `ST_Force3D` to put Z back
-       (clipped vertices land at z=0, which is acceptable for the few rows
-       that get clipped past ±85° lat).
-
-    fix(#888): returns the clip accounting — how many rows lost geometry
-    entirely (``dropped_features``) and how many survived in reduced form
-    (``clipped_features``) — so the caller can tell the user at the point of
-    loss instead of leaving them to hit "Analysis produced no features to
-    save" three steps later. Returns None when the table has no registered
-    ``geom`` metadata (nothing was inspected, let alone clipped).
-    """
-    _validate_table_name(table_name)
-    _validate_table_name(schema)
-
-    geom_meta = await session.execute(
-        text(
-            "SELECT srid, coord_dimension FROM geometry_columns "
-            "WHERE f_table_schema = :schema "
-            "  AND f_table_name = :table_name "
-            "  AND f_geometry_column = 'geom'"
-        ).bindparams(schema=schema, table_name=table_name)
-    )
-    row = geom_meta.first()
-    if row is None:
-        return None  # column has no registered metadata — nothing safe to clip
-    src_srid = int(row[0])
-    column_is_3d = int(row[1]) >= 3
-
-    shifted = await _shift_zero_to_360_longitudes(session, table_name, schema, src_srid)
-
-    if src_srid == 4326:
-        envelope = _MERCATOR_SAFE_ENVELOPE
-    else:
-        envelope = f"ST_Transform({_MERCATOR_SAFE_ENVELOPE}, {src_srid})"
-        # fix(#906): the guard runs AFTER the 0..360 shift above — that
-        # ordering is load-bearing (#888/#899): skipping the clip for a
-        # narrow-validity CRS must still leave the shift applied.
-        if await _mercator_envelope_degenerates(session, table_name, schema, src_srid):
-            logger.warning(
-                "Skipping the Web Mercator clip: the safe envelope degenerates "
-                "in the source CRS and the intersection would destroy data",
-                table=table_name,
-                schema=schema,
-                srid=src_srid,
-            )
-            return {
-                "shifted_longitudes": shifted,
-                "dropped_features": 0,
-                "clipped_features": 0,
-                "clip_skipped": True,
-            }
-
-    clipped = f"ST_CollectionExtract(ST_Intersection(geom, {envelope}), ST_Dimension(geom) + 1)"
-    if column_is_3d:
-        clipped = f"ST_Force3D({clipped})"
-
-    # fix(#888): count what the clip destroyed in the same statement that
-    # destroys it. Rows that were already empty are excluded from the WHERE so
-    # they cannot inflate the counts (the clip was a no-op for them anyway).
-    counts = (
-        await session.execute(
-            text(
-                f"WITH clip AS ("
-                f"  UPDATE {_qtable(table_name, schema=schema)} SET geom = {clipped} "
-                f"  WHERE geom IS NOT NULL AND NOT ST_IsEmpty(geom) "
-                f"    AND NOT ST_CoveredBy(geom, {envelope}) "
-                f"  RETURNING geom AS new_geom"
-                f") SELECT "
-                f"count(*) FILTER (WHERE new_geom IS NULL OR ST_IsEmpty(new_geom)), "
-                f"count(*) FILTER (WHERE new_geom IS NOT NULL AND NOT ST_IsEmpty(new_geom)) "
-                f"FROM clip"
-            )
-        )
-    ).first()
-    dropped_features = int(counts[0]) if counts is not None else 0
-    clipped_features = int(counts[1]) if counts is not None else 0
-    if dropped_features or clipped_features:
-        logger.warning(
-            "Geometry clipped to the Web Mercator safe envelope",
-            table=table_name,
-            schema=schema,
-            dropped_features=dropped_features,
-            clipped_features=clipped_features,
-            shifted_longitudes=shifted,
-        )
-    # ING-02 / P2-02 (Phase 1076): no internal commit. The caller
-    # (_finalize_ingest at tasks_common.py:821) owns the phase-2 commit
-    # boundary so a downstream failure rolls back this clip atomically.
-    return {
-        "shifted_longitudes": shifted,
-        "dropped_features": dropped_features,
-        "clipped_features": clipped_features,
-    }
-
-
-async def add_4326_column(
-    session: AsyncSession,
-    table_name: str,
-    source_srid: int,
-    *,
-    schema: str = "data",
-) -> None:
-    """Add a geom_4326 column with WGS84 geometry and spatial index.
-
-    If source_srid is 4326, copies geom directly (ensuring SRID is set).
-    Otherwise, reprojects via ST_Transform.
-
-    The 4326 column is declared 2D (`geometry(Geometry, 4326)`) — it backs
-    tile/map rendering, which is inherently 2D. If the source `geom` is 3D
-    (e.g. SRID 4979 with elevation), `ST_Force2D` strips Z so the UPDATE
-    doesn't fail with `Geometry has Z dimension but column does not`. Z is
-    still preserved in the original `geom` column.
-
-    fix(#1104): the column is also always LINEAR. WFS ingest admits curved
-    geometries (MultiSurface/CompoundCurve), and every surface that reads
-    geom_4326 raises on them: ST_AsMVTGeom (vector tiles), ST_AsGeoJSON
-    (feature reads), ``::geography`` and ST_MakeValid (analysis).
-    `ST_CurveToLine` densifies arcs here, at the one boundary they all read
-    from; it is an exact no-op on already-linear input, and the curved
-    source stays in the original `geom` column, same as Z does.
-    """
-    tref = _qtable(table_name, schema=schema)
-
-    await session.execute(
-        text(
-            f"ALTER TABLE {tref} "
-            f"ADD COLUMN IF NOT EXISTS geom_4326 geometry(Geometry, 4326)"
-        )
-    )
-
-    # fix(#1113 review r16): linearize IN THE SOURCE CRS, then reproject. An
-    # arc is defined by its control points, and CRS transforms are nonlinear:
-    # transforming the control points first and densifying after traces the
-    # arc in the wrong space, so ST_CurveToLine(ST_Transform(...)) yields a
-    # materially different shape from the correct
-    # ST_Transform(ST_CurveToLine(...)).
-    if source_srid == 4326:
-        rewrite_expr = "ST_Force2D(ST_CurveToLine(ST_SetSRID(geom, 4326)))"
-    else:
-        rewrite_expr = "ST_Force2D(ST_Transform(ST_CurveToLine(geom), 4326))"
-    await session.execute(text(f"UPDATE {tref} SET geom_4326 = {rewrite_expr}"))
-
-    await ensure_geom_4326_gist_index(session, table_name, schema=schema)
-
-    # DBM-05 (Phase 271): the previously-created `idx_<table>_gid` btree
-    # was redundant with the PK btree on `gid SERIAL PRIMARY KEY`. Removed
-    # so new ingests no longer ship the duplicate. Migration 0001_baseline drops
-    # the leftovers from existing tables.
-
-    # ING-02 / P2-02 (Phase 1076): no internal commit. The caller
-    # (_finalize_ingest at tasks_common.py:821) owns the phase-2 commit
-    # boundary so a downstream failure rolls back the ALTER + UPDATE +
-    # CREATE INDEX above atomically.
-
-
-async def linearize_existing_4326(
-    session: AsyncSession, table_name: str, *, schema: str = "data"
-) -> None:
-    """Enforce the geom_4326-is-always-linear invariant on a column we did not write.
-
-    fix(#1113 review): ``register_existing_table`` skips :func:`add_4326_column`
-    when the table already carries geom_4326, so a table created or copied into
-    the data schema AFTER migration 0034 ran could re-introduce curved values
-    the backfill can no longer see — and the per-read ST_CurveToLine wraps that
-    used to absorb them are gone. Registration is the app's write boundary for
-    such tables, so the invariant is enforced here, with the same predicate as
-    the migration: any arc, any top-level curve type, or any
-    GEOMETRYCOLLECTION (curve members cannot hide anywhere else — linear multi
-    types cannot contain them). Exact no-op on already-linear rows.
-
-    A BYO column may also DECLARE a curved typmod — geometry(CurvePolygon,
-    4326) — which would reject the linear UPDATE result outright; such a
-    column is loosened to a generic typmod first, PRESERVING its Z/M flags
-    (geometry_columns reports M as a type suffix and Z only via
-    coord_dimension, and a plain Geometry typmod rejects Z values). Only the
-    concrete curve typmods need it (abstract CURVE/SURFACE accept their
-    linear subtypes); rtrim(type,'M') matches the M-suffixed variants — no
-    base curve name ends in M.
-    """
-    tref = _qtable(table_name, schema=schema)
-    # fix(#1113 review r7): a STORED GENERATED geom_4326 rejects any UPDATE at
-    # parse time (even one whose WHERE matches nothing), and its values are
-    # decided by its generation expression, so it can be neither repaired nor
-    # safely retyped here — skip it (#1114 tracks expressions that yield
-    # curves).
-    generated = (
-        await session.execute(
-            text(
-                "SELECT 1 FROM information_schema.columns "
-                "WHERE table_schema = :schema AND table_name = :table "
-                "  AND column_name = 'geom_4326' AND is_generated = 'ALWAYS'"
-            ).bindparams(schema=schema, table=table_name)
-        )
-    ).first()
-    if generated is not None:
-        # fix(#1113 review r8): a generated column whose CURRENT rows are
-        # curved would register a dataset broken on every surface, and no
-        # later write of ours can fix it — refuse with the actionable cause
-        # instead. An empty or linear generated column registers fine; an
-        # expression that only yields curves for FUTURE rows is #1114's
-        # residue, same as any post-registration external write.
-        # fix(#1113 review r9): the test is "would linearization change the
-        # value", byte-for-byte — it catches arcs, top-level curve types, AND
-        # curve containers nested inside a GEOMETRYCOLLECTION with one
-        # comparison, while an all-linear collection (which ST_CurveToLine
-        # returns unchanged) stays registrable. A type list here would either
-        # miss the nested case or over-reject linear collections.
-        curved = (
-            await session.execute(
-                text(
-                    f"SELECT 1 FROM {tref} "  # noqa: S608
-                    f"WHERE ST_AsBinary(ST_CurveToLine(geom_4326)) "
-                    f"      <> ST_AsBinary(geom_4326) "
-                    f"LIMIT 1"
-                )
-            )
-        ).first()
-        if curved is not None:
-            raise ValueError(
-                "geom_4326 is a generated column whose expression yields "
-                "curved geometries; adjust it to apply ST_CurveToLine "
-                "(curved types break tiles, feature reads, and analysis)"
-            )
-        return
-    typmod = (
-        await session.execute(
-            text(
-                "SELECT type, srid, coord_dimension "
-                "FROM public.geometry_columns "
-                "WHERE f_table_schema = :schema "
-                "  AND f_table_name = :table "
-                "  AND f_geometry_column = 'geom_4326' "
-                "  AND rtrim(type, 'M') IN ('CIRCULARSTRING','COMPOUNDCURVE',"
-                "               'CURVEPOLYGON','MULTICURVE','MULTISURFACE')"
-            ).bindparams(schema=schema, table=table_name)
-        )
-    ).first()
-    if typmod is not None:
-        if typmod.coord_dimension == 4:
-            generic = "GeometryZM"
-        elif typmod.coord_dimension == 3:
-            generic = "GeometryM" if typmod.type.endswith("M") else "GeometryZ"
-        else:
-            generic = "Geometry"
-        await session.execute(
-            text(
-                f"ALTER TABLE {tref} ALTER COLUMN geom_4326 "
-                f"TYPE geometry({generic}, {int(typmod.srid)})"
-            )
-        )
-    # rtrim on GeometryType too: an M curve reports CURVEPOLYGONM, so the
-    # bare list would skip an arc-free M container.
-    await session.execute(
-        text(
-            f"UPDATE {tref} SET geom_4326 = ST_CurveToLine(geom_4326) "
-            f"WHERE ST_HasArc(geom_4326) "
-            f"   OR rtrim(GeometryType(geom_4326), 'M') IN "
-            f"      ('CIRCULARSTRING','COMPOUNDCURVE','CURVEPOLYGON',"
-            f"       'MULTICURVE','MULTISURFACE') "
-            f"   OR rtrim(GeometryType(geom_4326), 'M') = 'GEOMETRYCOLLECTION'"
-        )
-    )
-
-
-async def ensure_geom_4326_gist_index(
-    session: AsyncSession, table_name: str, *, schema: str = "data"
-) -> None:
-    """Create the GIST index on geom_4326 if this table doesn't have one.
-
-    fix(#448): the previous ``CREATE INDEX IF NOT EXISTS idx_<table>_geom_4326``
-    matched by NAME schema-wide, not per-table. On a second re-ingest the
-    previous swap's index (created against ``<table>_staging`` and carried
-    along by the RENAME) still held that name, so the new staging table
-    silently got NO spatial index — and the swap then dropped the only
-    indexed copy of the data. Check ``pg_indexes`` for a gist index on THIS
-    table instead, and let PostgreSQL pick a collision-free index name.
-    Called from both add_4326_column (staging load) and _apply_reupload_swap
-    (post-swap belt-and-braces), so any re-ingest self-heals a missing index.
-
-    The no-geom_4326 early return is defensive, not a reachable state (#1020):
-    add_4326_column has just added the column, and the swap call is gated on a
-    geometry_type that extract_metadata cannot report without reading geom_4326.
-    """
-    has_col = await session.execute(
-        text(
-            "SELECT 1 FROM information_schema.columns "
-            "WHERE table_schema = :schema AND table_name = :tn "
-            "AND column_name = 'geom_4326'"
-        ).bindparams(schema=schema, tn=table_name)
-    )
-    if has_col.first() is None:
-        return
-
-    has_gist = await session.execute(
-        text(
-            "SELECT 1 FROM pg_indexes "
-            "WHERE schemaname = :schema AND tablename = :tn "
-            "AND indexdef LIKE '%USING gist (geom_4326)%'"
-        ).bindparams(schema=schema, tn=table_name)
-    )
-    if has_gist.first() is None:
-        await session.execute(
-            text(
-                f"CREATE INDEX ON {_qtable(table_name, schema)} USING GIST (geom_4326)"
-            )
-        )
-
-
-async def grant_reader_access(
-    session: AsyncSession,
-    table_name: str,
-    *,
-    schema: str = "data",
-    role: str = "geolens_reader",
-) -> None:
-    """Grant SELECT on the table to the appropriate reader role.
-
-    DBM-12 (Phase 271): Kept as a defense-in-depth measure alongside
-    ``ALTER DEFAULT PRIVILEGES`` in ``scripts/init-db.sh``. If the runtime
-    ingest role matches the init-db role, this call is redundant; if they
-    differ (some custom deployment topologies), this is the only path that
-    grants SELECT on freshly-created tables.
-
-    In single_tenant: schema='data', role='geolens_reader' (unchanged behavior).
-    In multi_tenant: callers pass schema=tenant_data_schema(tid),
-                     role=tenant_reader_role(tid).
-
-    Parameters
-    ----------
-    session:
-        Active async SQLAlchemy session (caller controls the transaction).
-    table_name:
-        The table to GRANT SELECT on. Validated by _qtable.
-    schema:
-        Schema containing the table. Defaults to 'data' (single_tenant).
-    role:
-        Reader role to grant to. Defaults to 'geolens_reader' (single_tenant).
-    """
-    await session.execute(
-        text(f"GRANT SELECT ON {_qtable(table_name, schema)} TO {role}")
-    )
-    # ING-02 / P2-02 (Phase 1076): no internal commit. The caller
-    # (_finalize_ingest at tasks_common.py:821) owns the phase-2 commit
-    # boundary so a downstream failure rolls back this GRANT atomically.
-
-
-# ---------------------------------------------------------------------------
-# Attribute metadata helpers
-# ---------------------------------------------------------------------------
-
-
-def _humanize_column_name(field_name: str) -> str:
-    """Convert column name to human-readable title.
-
-    Examples:
-        pop_2020 -> Pop 2020
-        land_use_type -> Land Use Type
-        AREA_SQ_KM -> Area Sq Km
-        objectid -> Objectid
-        camelCaseField -> Camel Case Field
-    """
-    # Replace underscores with spaces
-    name = re.sub(r"_+", " ", field_name)
-    # Split camelCase boundaries
-    name = re.sub(r"([a-z])([A-Z])", r"\1 \2", name)
-    return name.strip().title()
-
-
-_UNIT_SUFFIX_MAP = {
-    "_m": "meters",
-    "_ft": "feet",
-    "_km": "kilometers",
-    "_mi": "miles",
-    "_sqm": "square meters",
-    "_sqft": "square feet",
-    "_sqkm": "square kilometers",
-    "_ha": "hectares",
-    "_ac": "acres",
-    "_pct": "percent",
-    "_deg": "degrees",
-    "_rad": "radians",
-    "_kg": "kilograms",
-    "_lb": "pounds",
-    "_l": "liters",
-    "_gal": "gallons",
-    "_s": "seconds",
-    "_min": "minutes",
-    "_hr": "hours",
-}
-
-
-def _infer_units(field_name: str) -> str | None:
-    """Infer units from column name suffix."""
-    lower = field_name.lower()
-    # Check longer suffixes first to avoid false matches (e.g. _sqm before _m)
-    for suffix, unit in sorted(_UNIT_SUFFIX_MAP.items(), key=lambda x: -len(x[0])):
-        if lower.endswith(suffix):
-            return unit
-    return None
-
-
-def _infer_semantic_role(field_name: str, data_type: str) -> str:
-    """Infer semantic role from column name and PostgreSQL data type."""
-    lower = field_name.lower()
-
-    # Geometry detection
-    if "geometry" in data_type.lower() or data_type == "USER-DEFINED":
-        return "geometry"
-
-    # Identifier patterns
-    if lower in ("id", "fid", "objectid", "gid", "ogc_fid") or lower.endswith("_id"):
-        return "identifier"
-
-    # Temporal patterns
-    if data_type in ("date", "timestamp without time zone", "timestamp with time zone"):
-        return "temporal"
-    if any(kw in lower for kw in ("date", "time", "year", "month", "day")):
-        return "temporal"
-
-    # Numeric -> measure
-    if data_type in (
-        "integer",
-        "bigint",
-        "smallint",
-        "numeric",
-        "double precision",
-        "real",
-    ):
-        return "measure"
-
-    # Label patterns
-    if lower in ("name", "label", "title", "display_name"):
-        return "label"
-
-    # Text -> categorical
-    if data_type in ("character varying", "text", "character"):
-        return "categorical"
-
-    return "other"
-
-
-_PG_TYPE_TO_DOMAIN = {
-    "integer": "discrete",
-    "bigint": "discrete",
-    "smallint": "discrete",
-    "numeric": "continuous",
-    "double precision": "continuous",
-    "real": "continuous",
-    "boolean": "boolean",
-    "date": "temporal",
-    "timestamp without time zone": "temporal",
-    "timestamp with time zone": "temporal",
-    "character varying": "categorical",
-    "character": "categorical",
-    "text": "text",
-    "USER-DEFINED": "geometry",
-    "ARRAY": "text",
-    "jsonb": "text",
-    "json": "text",
-    "uuid": "categorical",
-}
-
-
-def _infer_domain_type(data_type: str) -> str:
-    """Map PostgreSQL data_type to domain classification."""
-    return _PG_TYPE_TO_DOMAIN.get(data_type, "text")
-
-
-def _build_attribute_metadata(
-    AttributeMetadata: type,
-    dataset_id: uuid.UUID,
-    col_name: str,
-    col_type: str,
-    *,
-    sample_values: dict | None = None,
-    ordinal_position: int | None = None,
-    is_nullable: bool | None = None,
-) -> "Attribute":
-    """Factory for creating a new AttributeMetadata row with inferred fields.
-
-    Shared by generate_attribute_metadata (initial ingest) and
-    refresh_attribute_metadata (re-upload new columns). Callers resolve
-    the AttributeMetadata ORM class once via the Port and pass it in so
-    we don't re-do that lookup on every iteration of a per-column loop
-    (Phase 225 review fix W-03).
-    """
-    example_vals = None
-    if sample_values and col_name in sample_values:
-        example_vals = sample_values[col_name]
-
-    return AttributeMetadata(
-        dataset_id=dataset_id,
-        field_name=col_name,
-        title=_humanize_column_name(col_name),
-        data_type=col_type,
-        units=_infer_units(col_name),
-        semantic_role=_infer_semantic_role(col_name, col_type),
-        domain_type=_infer_domain_type(col_type),
-        example_values=example_vals,
-        ordinal_position=ordinal_position,
-        is_nullable=is_nullable,
-        is_current=True,
-    )
-
-
-def _build_geometry_attribute_row(
-    AttributeMetadata: type,
-    dataset_id: uuid.UUID,
-    geometry_type: str | None,
-) -> "Attribute":
-    """Factory for the special ``geom`` attribute metadata row.
-
-    Callers pass the resolved ORM class to avoid redundant Port lookups
-    in tight loops (Phase 225 review fix W-03).
-    """
-    return AttributeMetadata(
-        dataset_id=dataset_id,
-        field_name="geom",
-        title="Geometry",
-        data_type=geometry_type or "geometry",
-        semantic_role="geometry",
-        domain_type="geometry",
-        example_values=None,
-        is_current=True,
-    )
-
-
-async def generate_attribute_metadata(
-    session: AsyncSession,
-    dataset_id: uuid.UUID,
-    column_info: list[dict],
-    *,
-    geometry_type: str | None = None,
-    sample_values: dict | None = None,
-) -> list["Attribute"]:
-    """Auto-populate attribute_metadata rows from column_info.
-
-    Creates one row per column plus a geometry row if geometry_type is provided.
-    Uses check-then-insert to be idempotent (skips existing field_names).
-    Does NOT query the data table -- sample_values are passed in by the caller.
-    """
-    from app.platform.extensions import get_processing_port
-
-    AttributeMetadata = get_processing_port().get_attribute_metadata_orm_class()
-
-    # Load existing field names to skip duplicates
-    result = await session.execute(
-        select(AttributeMetadata.field_name).where(
-            AttributeMetadata.dataset_id == dataset_id
-        )
-    )
-    existing_fields = {row[0] for row in result.all()}
-
-    created: list[AttributeMetadata] = []
-
-    for col in column_info:
-        field_name = col["name"]
-        if field_name in existing_fields:
-            continue
-
-        am = _build_attribute_metadata(
-            AttributeMetadata,
-            dataset_id,
-            field_name,
-            col.get("type", ""),
-            sample_values=sample_values,
-            ordinal_position=col.get("ordinal_position"),
-            is_nullable=col.get("is_nullable"),
-        )
-        session.add(am)
-        created.append(am)
-        existing_fields.add(field_name)
-
-    # Geometry row
-    if geometry_type is not None and "geom" not in existing_fields:
-        am = _build_geometry_attribute_row(AttributeMetadata, dataset_id, geometry_type)
-        session.add(am)
-        created.append(am)
-
-    if created:
-        await session.flush()
-
-    return created
-
-
-async def refresh_attribute_metadata(
-    session: AsyncSession,
-    dataset_id: uuid.UUID,
-    column_info: list[dict],
-    *,
-    geometry_type: str | None = None,
-    sample_values: dict | None = None,
-) -> None:
-    """Refresh attribute metadata on re-upload, preserving user edits.
-
-    - Always refreshes system fields: data_type, example_values, ordinal_position,
-      is_nullable. Sets is_current=True.
-    - Per-field check: only refreshes title/semantic_role/domain_type/units/description
-      if that specific field name is NOT in user_modified_fields.
-    - New columns get auto-populated metadata.
-    - Removed columns are marked is_current=False.
-    """
-    from app.platform.extensions import get_processing_port
-
-    AttributeMetadata = get_processing_port().get_attribute_metadata_orm_class()
-
-    # Load existing attribute rows keyed by field_name
-    result = await session.execute(
-        select(AttributeMetadata).where(AttributeMetadata.dataset_id == dataset_id)
-    )
-    existing: dict[str, AttributeMetadata] = {
-        am.field_name: am for am in result.scalars().all()
-    }
-
-    current_field_names = {col["name"] for col in column_info}
-
-    for col in column_info:
-        field_name = col["name"]
-        data_type = col.get("type", "")
-
-        example_vals = None
-        if sample_values and field_name in sample_values:
-            example_vals = sample_values[field_name]
-
-        if field_name in existing:
-            am = existing[field_name]
-            # Always refresh system fields
-            am.data_type = data_type
-            am.example_values = example_vals
-            am.ordinal_position = col.get("ordinal_position")
-            am.is_nullable = col.get("is_nullable")
-            am.is_current = True
-
-            # Per-field check for user-editable fields
-            modified = set(am.user_modified_fields or [])
-            if "title" not in modified:
-                am.title = _humanize_column_name(field_name)
-            if "semantic_role" not in modified:
-                am.semantic_role = _infer_semantic_role(field_name, data_type)
-            if "domain_type" not in modified:
-                am.domain_type = _infer_domain_type(data_type)
-            if "units" not in modified:
-                am.units = _infer_units(field_name)
-            if "description" not in modified:
-                am.description = None  # No auto-inferred description
-        else:
-            # New column -- create fresh row via shared factory
-            am = _build_attribute_metadata(
-                AttributeMetadata,
-                dataset_id,
-                field_name,
-                data_type,
-                sample_values=sample_values,
-                ordinal_position=col.get("ordinal_position"),
-                is_nullable=col.get("is_nullable"),
-            )
-            session.add(am)
-
-    # Handle geometry row
-    if geometry_type is not None:
-        if "geom" in existing:
-            geom_am = existing["geom"]
-            geom_am.data_type = geometry_type or "geometry"
-            geom_am.is_current = True
-            modified = set(geom_am.user_modified_fields or [])
-            if "title" not in modified:
-                geom_am.title = "Geometry"
-            if "semantic_role" not in modified:
-                geom_am.semantic_role = "geometry"
-            if "domain_type" not in modified:
-                geom_am.domain_type = "geometry"
-        else:
-            session.add(
-                _build_geometry_attribute_row(
-                    AttributeMetadata, dataset_id, geometry_type
-                )
-            )
-
-    # Mark removed columns as is_current=False
-    for field_name, am in existing.items():
-        if field_name not in current_field_names and field_name != "geom":
-            am.is_current = False
-
-    await session.flush()
+__all__ = [
+    "_ABSTRACT_TO_CONCRETE_GEOMETRY_TYPE",
+    "_BOX3D_RE",
+    "_DEGREE_UNIT_SRTEXT_RE",
+    "_GEOGRAPHIC_SRTEXT_RE",
+    "_MERCATOR_SAFE_ENVELOPE",
+    "_PG_TYPE_TO_DOMAIN",
+    "_TABLE_NAME_RE",
+    "_UNIT_SUFFIX_MAP",
+    "_build_attribute_metadata",
+    "_build_geometry_attribute_row",
+    "_humanize_column_name",
+    "_infer_domain_type",
+    "_infer_semantic_role",
+    "_infer_units",
+    "_mercator_envelope_degenerates",
+    "_normalize_geometry_type",
+    "_parse_box3d_z_bounds",
+    "_qtable",
+    "_score_attribute_completeness",
+    "_score_crs",
+    "_score_geometry_validity",
+    "_score_metadata_completeness",
+    "_seam_crossing_extent_wkt",
+    "_shift_zero_to_360_longitudes",
+    "_sql_quote_ident",
+    "_table_has_geometry",
+    "_validate_table_name",
+    "add_4326_column",
+    "clip_to_mercator_bounds",
+    "compute_quality_score",
+    "construct_point_geometry",
+    "construct_wkt_geometry",
+    "detect_3d_metadata",
+    "detect_dbf_truncation_collisions",
+    "ensure_geom_4326_gist_index",
+    "ensure_geom_column",
+    "extract_metadata",
+    "generate_attribute_metadata",
+    "get_column_info",
+    "get_extent",
+    "get_feature_count",
+    "get_geometry_type",
+    "get_sample_values",
+    "get_table_srid",
+    "grant_reader_access",
+    "linearize_existing_4326",
+    "promote_z_to_elev",
+    "refresh_attribute_metadata",
+    "rename_reserved_columns",
+]
